Added padding to input range
bug: Fixes issue with sun animation clipping.

The sun's final position needed to be set on the base styles for the
sun itself and not just in the keyframe. This caused an issue where
the sun would use the default position behavior after the animation
completed.

refactor: Uses div tag for sun element.

div is a slightly more semantically correct tag for the sun than p.

bug: Fixed overflow issue on mobile.

bug: Webpage overflowed on mobile. I set the overflow style to hidden on the \'root\' dive element.